requirements.txt
django~=4.2
torch
transformers
requests
deepl
nltk
openai
python-dotenv

